+++ lib/components/dependency_view.dart
