+++ lib/components/dependency_view.html
@@ -4,7 +4,7 @@ for details. All rights reserved. Use of this source code is governed by a
 BSD-style license that can be found in the LICENSE file.
 -->
 
-<link rel="import" href="../packages/polymer/polymer.html">
+<link rel="import" href="../../../packages/polymer/polymer.html">
 <polymer-element name="dependency-view">
   <template>
     <style>
@@ -49,7 +49,7 @@ BSD-style license that can be found in the LICENSE file.
     </style>
     <div id="information">
       In order to use the dependency view, first find an element to focus on
-      in the hierarchy view and click on the <img src="../deps_icon.svg"> symbol.
+      in the hierarchy view and click on the <img src="deps_icon.svg"> symbol.
     </div>
     <div id="tables" style="display:none;">
       <table id="in">
