+++ lib/components/diff_view.html
@@ -4,7 +4,7 @@ for details. All rights reserved. Use of this source code is governed by a
 BSD-style license that can be found in the LICENSE file.
 -->
 
-<link rel="import" href="../packages/polymer/polymer.html">
+<link rel="import" href="../../../packages/polymer/polymer.html">
 <polymer-element name="diff-view">
     <template>
         <style>
