+++ lib/dragdrop.dart
@@ -2,7 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of viewer;
+library dump_viz.dragdrop;
+
+import 'dart:async';
+import 'dart:html';
 
 class DragDropFile {
   Element _dropZone;
