+++ lib/history.dart
@@ -2,11 +2,12 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-library history;
+library dump_viz.history;
 
-import 'dart:html';
 import 'dart:async';
-import 'polymer_lib/dependency_view.dart';
+import 'dart:html';
+
+import 'components/dependency_view.dart';
 
 abstract class HistoryState {
   /// Apply this history state and modify the view.
