+++ lib/infohelper.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+library dump_viz.info_helper;
+
 import 'dart:collection' show Queue;
 
 class Selection {
