+++ lib/src/versions.dart
@@ -6,11 +6,10 @@ library versions;
 
 import 'dart:html';
 
-import 'package:dump_viz/infohelper.dart';
-
-import '../polymer_lib/tree_table.dart';
-import '../polymer_lib/dependency_view.dart';
+import '../components/dependency_view.dart';
+import '../components/tree_table.dart';
 import '../history.dart';
+import '../infohelper.dart';
 
 part 'version0.dart';
 part 'version1.dart';
