+++ web/index.html
@@ -12,14 +12,14 @@ BSD-style license that can be found in the LICENSE file.
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="viewer.css">
 
-    <link rel="import" href="packages/paper_elements/roboto.html">
-    <link rel="import" href="polymer_lib/tree_table.html">
-    <link rel="import" href="polymer_lib/dependency_view.html">
-    <link rel="import" href="polymer_lib/diff_view.html">
-
     <link rel="import" href="packages/core_elements/core_toolbar.html">
     <link rel="import" href="packages/paper_elements/paper_tab.html">
     <link rel="import" href="packages/paper_elements/paper_tabs.html">
+    <link rel="import" href="packages/paper_elements/roboto.html">
+
+    <link rel="import" href="packages/dump_viz/components/dependency_view.html">
+    <link rel="import" href="packages/dump_viz/components/diff_view.html">
+    <link rel="import" href="packages/dump_viz/components/tree_table.html">
   </head>
   <body unresolved>
     <core-toolbar>
